banana skull upd, blunderbuss, pp & pr mats
Updated the bandana skull to be more accurate to its game counterpart, but now that it is infinite ammo with reloading, i removed the damage penalty, but kept the recoil and other drawbacks.
Added the blunderbuss, along with an integrated perk. Its just a better airbender in tl;dr.
Added the VMTs for the CE plasma rifle and plasma pistol, though i have not yet set up the skins.